Fix: first round of testing and fixes with new
