+++ src/sparseml/pytorch/sparsification/pruning/scorer.py
@@ -105,7 +105,7 @@ class PruningParamsGradScorer(PruningParamsScorer, ABC):
     def __init__(
             self, 
             params: List[Parameter],
-            grad_sampler: Optional(GradSampler),
+            grad_sampler: Optional[GradSampler],
         ):
         super().__init__(params=params)
