+++ src/sparseml/pytorch/utils/sparsification.py
@@ -39,7 +39,10 @@ from sparseml.pytorch.utils.helpers import (
 )
 
 
-__all__ = ["ModuleSparsificationInfo"]
+__all__ = [
+    "ModuleSparsificationInfo",
+    "GradSampler",
+    ]
 
 
 class ModuleSparsificationInfo:
