+++ tests/sparseml/pytorch/sparsification/pruning/test_mfac.py
@@ -17,7 +17,7 @@ import os
 import pytest
 import torch
 
-from sparseml.sparsification import (
+from sparseml.pytorch.sparsification.pruning import (
     FisherInverseFast,
     FisherInverseFastBlock,
     FisherInverseFastSmallBlocks,
@@ -60,7 +60,10 @@ def test_fast_small_blocks():
     small_blocks_mul_out = small_blocks_hinv.mul(tensor_to_mul)
 
     assert (
-        pytest.approx(fast_diag, torch.mean(fast_diag).item() * PRECISION)
+        pytest.approx(
+            fast_diag, 
+            torch.sqrt(torch.mean(torch.square(fast_diag))).item() * PRECISION
+        ) 
         == small_blocks_diag
     )
     assert (
@@ -102,15 +105,17 @@ def test_fast_small_blocks_gpu():
     small_blocks_mul_out = small_blocks_hinv.mul(tensor_to_mul)
 
     assert (
-        pytest.approx(fast_diag, torch.mean(fast_diag).item() * PRECISION)
+        pytest.approx(
+            fast_diag, 
+            torch.sqrt(torch.mean(torch.square(fast_diag))).item() * PRECISION
+        )
         == small_blocks_diag
     )
     assert (
         pytest.approx(
             fast_mul_out,
-            torch.mean(fast_mul_out).item() * PRECISION,
-        )
-        == small_blocks_mul_out
+            torch.sqrt(torch.mean(torch.square(fast_mul_out))).item() * PRECISION,
+        ) == small_blocks_mul_out
     )
 
 
@@ -146,13 +151,16 @@ def test_fast_large_blocks():
     small_blocks_mul_out = small_blocks_hinv.mul(tensor_to_mul)
 
     assert (
-        pytest.approx(fast_diag, torch.mean(fast_diag).item() * PRECISION)
+        pytest.approx(
+            fast_diag, 
+            torch.sqrt(torch.mean(torch.square(fast_diag))).item() * PRECISION
+        )
         == small_blocks_diag
     )
     assert (
         pytest.approx(
             fast_mul_out,
-            torch.mean(fast_mul_out).item() * PRECISION,
+            torch.sqrt(torch.mean(torch.square(fast_mul_out))).item() * PRECISION
         )
         == small_blocks_mul_out
     )
@@ -188,13 +196,16 @@ def test_fast_large_blocks_gpu():
     small_blocks_mul_out = small_blocks_hinv.mul(tensor_to_mul)
 
     assert (
-        pytest.approx(fast_diag, torch.mean(fast_diag).item() * PRECISION)
+        pytest.approx(
+            fast_diag, 
+            torch.sqrt(torch.mean(torch.square(fast_diag))).item() * PRECISION
+        )
         == small_blocks_diag
     )
     assert (
         pytest.approx(
             fast_mul_out,
-            torch.mean(fast_mul_out).item() * PRECISION,
+            torch.sqrt(torch.mean(torch.square(fast_mul_out))).item() * PRECISION
         )
         == small_blocks_mul_out
     )
